Move Footer component and set fixed position
